+++ packages/spacecat-shared-utils/test/llmo-config.test.js
@@ -71,6 +71,9 @@ describe('llmo-config utilities', () => {
         competitors: {
           competitors: [],
         },
+        deleted: {
+          prompts: {},
+        },
       });
     });
   });